Updated shops to a place on the map

+++ HealthSystem.cs
@@ -10,9 +10,9 @@ namespace WalkOfLegends
     {
 
 
-        public int health;
+        public float health;
 
-        public HealthSystem(int health)
+        public HealthSystem(float health)
         {
             this.health = health;
         }
